Finds the NN in the training set for a observation in the test set with a KNN index matrix

Takes an pre-searched index matrix and finds the correct nearest nearest neighbors for each bootstrap iteration. In the unbootstrapped run, the keys_standard will be the same as the keys. In a bootstrapped run, the keys parameter will contained the bootstrapped keys

bootstrapIndexMatrixProcessorTest(
  keys_standard,
  index_matrix,
  keys,
  keys_test,
  k_value
)

Arguments

keys_standard

A list of the unbootstrapped outcomes (training set) and keys (training and test sets) (list)

index_matrix

A matrix of indices of the NN (matrix)

keys

A vector of the keys in the training set (integer)

keys_test

A vector of the keys in the test set (integer)

k_value

A value for K for the causal KNN (integer)

Value

A vector of KNN estimates (numeric)